LayoutTests/platform/glib/imported/w3c/web-platform-tests/cors/request-headers-expected.txt

CONSOLE MESSAGE: Request header field y-print is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors/resources/cors-makeheader.py?headers=x-print due to access control checks.
Request headers


PASS basic request header
PASS Simple request headers need not be in allow-headers
PASS Unspecified request headers are disallowed
PASS Strange allowheaders (case insensitive)
PASS INVALID_STATE_ERR on setRequestHeader before open()
PASS INVALID_STATE_ERR on setRequestHeader after send()
